requirements.txt
cryptography==1.3.2
Django==1.9.4
djangorestframework==3.3.3